+++ Clicked/Core/Clicked.lua
@@ -131,6 +131,7 @@ function Clicked:OnEnable()
 	if WOW_PROJECT_ID == WOW_PROJECT_MAINLINE then
 		self:RegisterEvent("PLAYER_TALENT_UPDATE", "ReloadActiveBindings")
 		self:RegisterEvent("PLAYER_FLAGS_CHANGED", OnPlayerFlagsChanged)
+		self:RegisterEvent("PLAYER_FOCUS_CHANGED", "RefreshTooltips")
 	end
 
 	self:RegisterEvent("PLAYER_LEVEL_CHANGED", "ReloadActiveBindings");
@@ -139,7 +140,6 @@ function Clicked:OnEnable()
 
 	self:RegisterEvent("MODIFIER_STATE_CHANGED", "RefreshTooltips")
 	self:RegisterEvent("UNIT_TARGET", "RefreshTooltips")
-	self:RegisterEvent("PLAYER_FOCUS_CHANGED", "RefreshTooltips")
 
 	for _, module in pairs(modules) do
 		if module.Register ~= nil then
@@ -156,6 +156,7 @@ function Clicked:OnDisable()
 	if WOW_PROJECT_ID == WOW_PROJECT_MAINLINE then
 		self:UnregisterEvent("PLAYER_TALENT_UPDATE")
 		self:UnregisterEvent("PLAYER_FLAGS_CHANGED")
+		self:UnregisterEvent("PLAYER_FOCUS_CHANGED")
 	end
 
 	self:UnregisterEvent("PLAYER_LEVEL_CHANGED")
@@ -164,7 +165,6 @@ function Clicked:OnDisable()
 
 	self:UnregisterEvent("MODIFIER_STATE_CHANGED")
 	self:UnregisterEvent("UNIT_TARGET")
-	self:UnregisterEvent("PLAYER_FOCUS_CHANGED")
 
 	for _, module in pairs(modules) do
 		if module.Unregister ~= nil then