+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [0.16.2] - 2021-02-20
+
+### Fixed
+
+- Fix an error when logging in on Classic
+
 ## [0.16.1] - 2021-02-20
 
 ### Added
@@ -548,7 +554,8 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 - Initial public release
 
-[Unreleased]: https://github.com/Snakybo/Clicked/compare/0.16.1...master
+[Unreleased]: https://github.com/Snakybo/Clicked/compare/0.16.2...master
+[0.16.2]: https://github.com/Snakybo/Clicked/releases/tag/0.16.2
 [0.16.1]: https://github.com/Snakybo/Clicked/releases/tag/0.16.1
 [0.16.0]: https://github.com/Snakybo/Clicked/releases/tag/0.16.0
 [0.15.5]: https://github.com/Snakybo/Clicked/releases/tag/0.15.5